Revert "cluster-normal layout"
This reverts commit 307410f48d9c5f7f3fc932384bc884aac03f5a44.
